Rabi animations finished for Brightspace
